Improve tests and phpdoc.
